+++ geopandas/tests/test_array.py
@@ -1,4 +1,5 @@
 import random
+import warnings
 
 import numpy as np
 import pandas as pd
@@ -162,6 +163,26 @@ def test_from_wkb_hex():
     assert isinstance(res, GeometryArray)
 
 
+@pytest.mark.parametrize("on_invalid", ["raise", "warn", "ignore"])
+def test_from_wkb_on_invalid(on_invalid):
+    # Single point LineString hex WKB: invalid
+    invalid_wkb_hex = "01020000000100000000000000000008400000000000000840"
+    message = "point array must contain 0 or >1 elements"
+
+    if on_invalid == "raise":
+        with pytest.raises(Exception, match=message):
+            from_wkb([invalid_wkb_hex], on_invalid=on_invalid)
+    elif on_invalid == "warn":
+        with pytest.warns(Warning, match=message):
+            res = from_wkb([invalid_wkb_hex], on_invalid=on_invalid)
+            assert res == [None]
+    elif on_invalid == "ignore":
+        with warnings.catch_warnings():
+            warnings.simplefilter("error")
+            res = from_wkb([invalid_wkb_hex], on_invalid=on_invalid)
+            assert res == [None]
+
+
 def test_to_wkb():
     P = from_shapely(points_no_missing)
     res = to_wkb(P)
@@ -215,6 +236,26 @@ def test_from_wkt(string_type):
     )
     res = from_wkt([f(multi_poly.wkt)])
     assert res[0] == multi_poly
+
+
+@pytest.mark.parametrize("on_invalid", ["raise", "warn", "ignore"])
+def test_from_wkt_on_invalid(on_invalid):
+    # Single point LineString WKT: invalid
+    invalid_wkt = "LINESTRING(0 0)"
+    message = "point array must contain 0 or >1 elements"
+
+    if on_invalid == "raise":
+        with pytest.raises(Exception, match=message):
+            from_wkt([invalid_wkt], on_invalid=on_invalid)
+    elif on_invalid == "warn":
+        with pytest.warns(Warning, match=message):
+            res = from_wkt([invalid_wkt], on_invalid=on_invalid)
+            assert res == [None]
+    elif on_invalid == "ignore":
+        with warnings.catch_warnings():
+            warnings.simplefilter("error")
+            res = from_wkt([invalid_wkt], on_invalid=on_invalid)
+            assert res == [None]
 
 
 def test_to_wkt():