+++ CHANGELOG.md
@@ -17,6 +17,7 @@ New features and improvements:
 
 - GeoSeries and GeoDataFrame `__repr__` now trims trailing zeros for a more readable
   output (#3087).
+- Add `on_invalid` parameter to `from_wkt` and `from_wkb` (#)
 
 Potentially breaking changes:
 - reading a data source that does not have a geometry field using ``read_file``